requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
gradio>=5.0.0
python-multipart>=0.0.12
unstructured>=0.16.0
pymupdf>=1.24.0
python-docx>=1.1.0
easyocr>=1.7.0
Pillow>=10.0.0
sentence-transformers>=3.3.0
torch>=2.5.0
chromadb>=0.6.0
rank-bm25>=0.2.2
langchain>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
llama-index-core>=0.11.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-huggingface>=0.3.0
ragas>=0.1.0
datasets>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
pytest-cov>=5.0.0
